platform/backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.7,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
celery>=5.4,<6.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<26.0
uvicorn>=0.30,<1.0
prometheus-client>=0.21,<1.0
